+++ src/mine/stats.hpp
@@ -60,6 +60,7 @@ class Stats {
   std::map<Number, int64_t> num_constants;
   std::map<Operation, int64_t> num_operations;
   std::map<OpPos, int64_t> num_operation_positions;
+  std::map<std::string, int64_t> num_programs_per_submitter;
   std::multimap<UID, UID> call_graph;
   std::vector<int64_t> num_programs_per_length;
   std::vector<int64_t> num_ops_per_type;